+++ Submodules/AppNetwork/Sources/AppNetwork/Enums/Errors/SocketEnvironmentError.swift
